Removed string.FirstCharUpper extension method.
