+++ src/RhinoInside.Revit.GH/Extensions/System.cs
@@ -5,18 +5,6 @@ namespace RhinoInside.Revit
 {
   static partial class Extension
   {
-    #region string
-    public static string FirstCharUpper(this string text)
-    {
-      if (char.IsUpper(text, 0))
-        return text;
-
-      var chars = text.ToCharArray();
-      chars[0] = char.ToUpperInvariant(chars[0]);
-      return new string(chars);
-    }
-    #endregion
-
     #region Linq
     public static IEnumerable<K> Select<K, T>(this IEnumerator<T> e, Func<T, K> selector)
     {
